feat(agent): separar implementación de artwork por plataforma
- Linux: usa CEF API para soportar WebP/GIF animados
- Windows: usa filesystem directo (sin CEF, más rápido)
- Simplifica control_windows.go sin intentar CEF
- Hub filtra formatos a PNG/JPEG para Windows
- Actualiza submodule con SetArtworkFilesystem()
